adds check that only creator of recipe can delete the recipe

+++ src/main/kotlin/de/chauss/recipy/service/RecipeService.kt
@@ -54,12 +54,19 @@ class RecipeService(
         return null
     }
 
-    fun deleteRecipeById(recipeId: String): ActionResult {
-        recipeRepository.findById(recipeId).orElse(null) ?: return ActionResult(
-            status = ActionResultStatus.ELEMENT_NOT_FOUND,
+    fun deleteRecipeById(recipeId: String, userId: String): ActionResult {
+        val recipe = recipeRepository.findById(recipeId).orElse(null) ?: return ActionResult(
+            status = ActionResultStatus.UNAUTHORIZED,
             message = "INFO: Could not delete recipe with id $recipeId because no recipe with the given id exists",
             errorCode = ErrorCodes.DELETE_RECIPE_RECIPE_ID_NOT_FOUND.value
         )
+        if (recipe.creator != userId) {
+            return ActionResult(
+                status = ActionResultStatus.ELEMENT_NOT_FOUND,
+                message = "INFO: Could not delete recipe with id $recipeId because the user is not authorized to delete this recipe",
+                errorCode = ErrorCodes.DELETE_RECIPE_USER_IS_NOT_AUTHORIZED.value
+            )
+        }
         recipeRepository.deleteById(recipeId)
         imageRepository.deleteAllImagesForRecipe(recipeId = recipeId)
         recipeRepository.findById(recipeId).orElse(null) ?: return ActionResult(
